Properly build hoisted scripts
