+++ packages/astro/src/core/util.ts
@@ -3,7 +3,6 @@ import path from 'node:path';
 import { fileURLToPath } from 'node:url';
 import { normalizePath } from 'vite';
 import type { AstroConfig, AstroSettings, RouteType } from '../@types/astro.js';
-import { isServerLikeOutput } from '../prerender/utils.js';
 import { SUPPORTED_MARKDOWN_FILE_EXTENSIONS } from './constants.js';
 import type { ModuleLoader } from './module-loader/index.js';
 import { prependForwardSlash, removeTrailingForwardSlash, slash } from './path.js';
@@ -148,8 +147,19 @@ export function isEndpoint(file: URL, settings: AstroSettings): boolean {
 	return !endsWithPageExt(file, settings);
 }
 
+export function isServerLikeOutput(config: AstroConfig) {
+	return config.output === 'server' || config.output === 'hybrid';
+}
+
 export function isModeServerWithNoAdapter(settings: AstroSettings): boolean {
 	return isServerLikeOutput(settings.config) && !settings.adapter;
+}
+
+
+export function isContentCollectionsCacheEnabled(config: AstroConfig): boolean {
+	return config.experimental.contentCollectionCache &&
+		// contentCollectionsCache is an SSG only feature
+		!isServerLikeOutput(config);
 }
 
 export function relativeToSrcDir(config: AstroConfig, idOrUrl: URL | string) {
