+++ packages/astro/src/vite-plugin-scanner/index.ts
@@ -4,8 +4,8 @@ import type { Plugin as VitePlugin } from 'vite';
 import { normalizePath } from 'vite';
 import type { AstroSettings } from '../@types/astro.js';
 import { type Logger } from '../core/logger/core.js';
-import { isEndpoint, isPage, rootRelativePath } from '../core/util.js';
-import { getPrerenderDefault, isServerLikeOutput } from '../prerender/utils.js';
+import { isEndpoint, isPage, rootRelativePath, isServerLikeOutput } from '../core/util.js';
+import { getPrerenderDefault } from '../prerender/utils.js';
 import { scan } from './scan.js';
 
 export interface AstroPluginScannerOptions {
